fix(media): address bundle persistence review feedback
